Source code restructure
* `src/energomera_hass_mqtt`: Has been restructured to have classes in
  separate files (with a few exceptions)
+ `src/energomera_hass_mqtt`: Added support to specify port of MQTT
  broker (`mqtt.port` configuration file option). Also, `async_main()`
  now receives optional `port` argument to allow port being overridden
  at run time, solely used by tests (see next commit)
